reorganize hot reload module
